fix: remove redundant form-control class for MasqueradeUserInput
As the FormControl element already comes with the form-control
class, the extra class specified on the MasqueradeUserInput is
applied to the div container of the <input> causing double input
box effect. This is not visible in themes where the container
div's padding is zero. But appears when padding is not set.

Internal-ref: https://tasks.opencraft.com/browse/BB-8929